Correct previous checkin -- it would result in erroneous calculations when downsampling due to integer roundoff.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10486 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/src/mp/MprFromFile.cpp
@@ -614,7 +614,8 @@ UtlBoolean MprFromFile::allocateAndResample(char*& audBuf,
    }
 
    // Malloc up a new chunk of memory for resampling to.
-   uint32_t outBufSize = audBufSz * outRate/inRate;
+   uint32_t outBufSize = 
+      (uint32_t)(((uint64_t)audBufSz * (uint64_t)outRate) / (uint64_t)inRate);
    MpAudioSample* pOutSamples = (MpAudioSample*)malloc(outBufSize);
    if(pOutSamples == NULL)
    {
